+++ src/types.ts
@@ -1,5 +1,9 @@
 import { AxiosResponseHeaders } from "axios";
 
+export interface Paging {
+    next: string;
+}
+
 export interface Web3PagedResponse<T> {
     data: T[];
     paging?: Paging;
@@ -502,6 +506,81 @@ export interface TravelRuleVaspFilter {
     order?: string;
 }
 
+export interface ScreeningPolicyConfiguration {
+    bypassScreeningDuringServiceOutages?: boolean;
+    inboundTransactionDelay?: number;
+    outboundTransactionDelay?: number;
+}
+
+export enum TravelRuleAction {
+    screen = "SCREEN",
+    pass = "PASS",
+    freeze = "FREEZE"
+}
+
+export interface TravelRulePolicyRule {
+    sourceType?: string;
+    sourceSubType?: string;
+    destType?: string;
+    destSubType?: string;
+    destAddress?: string;
+    sourceId?: string;
+    destId?: string;
+    asset?: string;
+    baseAsset?: string;
+    amount?: number;
+    amountUSD?: number;
+    networkProtocol?: string;
+    operation?: string;
+    action: TravelRuleAction;
+}
+
+export enum PolicyApprovalStatus {
+    live = "live",
+    processing = "processing"
+}
+
+export enum TransactionDirection {
+    inbound = "INBOUND",
+    outbound = "OUTBOUND"
+}
+
+export enum FbTravelRuleTransactionStatus {
+    completed = "COMPLETED",
+    pending = "PENDING",
+    rejected = "REJECTED",
+    failed = "FAILED",
+    canceled = "CANCELED",
+    blockingTimeExpired = "BLOCKING_TIME_EXPIRED",
+}
+
+export enum TravelRuleVerdict {
+    accept = "ACCEPT",
+    reject = "REJECT",
+    alert = "ALERT",
+    wait = "WAIT",
+    freeze = "FREEZE",
+    cancel = "CANCEL"
+}
+
+export interface TravelRuleRulesConfiguration {
+    direction?: TransactionDirection;
+    status?: FbTravelRuleTransactionStatus;
+    amountUSD?: number;
+    amount?: number;
+    asset?: string;
+    action: TravelRuleVerdict;
+}
+
+export interface TravelRulePolicy {
+    tenantId?: string;
+    policy: TravelRulePolicyRule[];
+    policyStatus?: PolicyApprovalStatus;
+    isDefault: boolean;
+    createDate?: Date;
+    lastUpdate: Date;
+}
+
 export enum Web3ConnectionFeeLevel {
     HIGH = "HIGH",
     MEDIUM = "MEDIUM",
@@ -520,6 +599,16 @@ export interface ExchangeResponse {
     assets: AssetResponse[];
     isSubaccount: boolean;
     status: string;
+}
+
+export interface PagedExchangeResponse {
+    exchanges: ExchangeResponse[];
+    paging?: {
+        before?: string;
+        after?: string;
+    };
+    prevUrl?: string;
+    nextUrl?: string;
 }
 
 export interface ConvertExchangeAssetResponse {
@@ -763,6 +852,7 @@ export interface TransactionFilter {
     destType?: PeerType;
     sourceId?: string;
     destId?: string;
+    sort?: "ASC" | "DESC";
 }
 
 export interface NFTOwnershipFilter {
@@ -811,19 +901,15 @@ export interface GetNFTsFilter {
     order?: OrderValues;
 }
 
-class MediaEntity {
-    url: string;
-    contentType: string;
-}
-
 interface NFTCollection {
     id: string;
-    name: string;
-    symbol: string;
+    name?: string;
+    symbol?: string;
 }
 
-export interface Paging {
-    next: string;
+interface MediaEntity {
+    url: string;
+    contentType: string;
 }
 
 export interface Token {
@@ -831,9 +917,9 @@ export interface Token {
     tokenId: string;
     standard: string;
     blockchainDescriptor: string;
-    description: string;
-    name: string;
-    media: MediaEntity[];
+    description?: string;
+    name?: string;
+    media?: MediaEntity[];
     metadataURI?: string;
     cachedMetadataURI?: string;
     collection?: NFTCollection;
@@ -864,6 +950,13 @@ export interface TransactionPageFilter {
     destType?: PeerType;
     sourceId?: string;
     destId?: string;
+    sort?: "ASC" | "DESC";
+}
+
+export interface ExchangeAccountsPageFilter {
+    limit: number;
+    before?: string;
+    after?: string;
 }
 
 export enum TransactionOrder {
@@ -1400,6 +1493,11 @@ export enum NFTOwnershipStatus {
     "ARCHIVED" = "ARCHIVED",
 }
 
+export interface NFTOwnershipStatusUpdatedPayload {
+    assetId: string;
+    status: NFTOwnershipStatus;
+}
+
 export enum NFTOwnershipWalletType {
     "VAULT_ACCOUNT" = "VAULT_ACCOUNT",
     "END_USER_WALLET" = "END_USER_WALLET",
@@ -1453,27 +1551,44 @@ interface VendorDto {
     attributes: Record<string, string>;
 }
 
-export interface ContractTemplateDto {
-    isPublic: boolean;
-    vendor?: VendorDto;
+export interface ContractDeployRequest {
+    assetId: string;
+    vaultAccountId: string;
+    constructorParameters?: object[];
+}
+
+export interface ContractDeployResponse {
+    txId: string;
+    assetId: string;
+    vaultAccountId: string;
+    contractId: string;
+}
+
+export interface LeanContractTemplateDto {
     id: string;
     name: string;
     description: string;
-    bytecode: string;
-    sourcecode?: string;
+    attributes?: Record<string, string>;
+    isPublic: boolean;
     owner?: string;
+    vendor?: VendorDto;
+}
+
+export interface ContractTemplateDto {
+    id: string;
+    name: string;
+    description: string;
     compilerOutputMetadata?: object;
     abi: AbiFunction[];
-    docs?: ContractDoc;
     attributes?: Record<string, string>;
-}
-export enum TokenLinkPermission {
-    MINT = "MINT",
-    BURN = "BURN",
+    docs?: ContractDoc;
+    owner?: string;
+    vendor?: VendorDto;
+    isPublic: boolean;
 }
 
 export interface TokenLinkPermissionEntry {
-    permission: TokenLinkPermission;
+    permission: "MINT" | "BURN";
     vaultAccountId: string;
 }
 
@@ -1503,6 +1618,7 @@ export interface PendingTokenLinkDto {
     blockchainId?: string;
 }
 
+type CreateTokenParams = EVMTokenCreateParamsDto | StellarRippleCreateParamsDto;
 
 export interface IssueTokenRequest {
     symbol: string;
@@ -1549,6 +1665,7 @@ export class BatchTask {
 }
 
 type CreateTokenParams = EVMTokenCreateParamsDto | StellarRippleCreateParamsDto;
+
 interface StellarRippleCreateParamsDto {
     issuerAddress?: string;
 }
@@ -1702,6 +1819,12 @@ export namespace NCW {
         order?: "ASC" | "DESC";
     }
 
+    export interface GetSupportedAssetsPayload {
+        pageCursor?: string;
+        pageSize?: number;
+        onlyBaseAssets?: boolean;
+    }
+
     export interface GetWalletAccountsPayload {
         pageCursor?: string;
         pageSize?: number;
@@ -1758,5 +1881,190 @@ export namespace NCW {
     export interface Device {
         deviceId: string;
         enabled: boolean;
+    }
+}
+
+export namespace TAP {
+    type PolicyTransactionType =
+        | "*"
+        | "CONTRACT_CALL"
+        | "RAW"
+        | "TRANSFER"
+        | "APPROVE"
+        | "MINT"
+        | "BURN"
+        | "SUPPLY"
+        | "REDEEM"
+        | "STAKE"
+        | "TYPED_MESSAGE";
+
+    type PolicySrcOrDestType =
+        | "EXCHANGE"
+        | "UNMANAGED"
+        | "VAULT"
+        | "NETWORK_CONNECTION"
+        | "COMPOUND"
+        | "FIAT_ACCOUNT"
+        | "ONE_TIME_ADDRESS"
+        | "*";
+
+    type PolicyType = "TRANSFER";
+
+    type PolicyAction = "ALLOW" | "BLOCK" | "2-TIER";
+
+    type PolicyDestAddressType = "*" | "WHITELISTED" | "ONE_TIME";
+
+    type PolicyAmountScope = "SINGLE_TX" | "TIMEFRAME";
+
+    type PolicySrcOrDestSubType = "*" | "EXTERNAL" | "INTERNAL" | "CONTRACT" | "EXCHANGETEST";
+
+    type PolicySrcOrDestId = string;
+
+    type AuthorizationGroup = {
+        users?: Array<string>;
+        usersGroups?: Array<string>;
+        th: number;
+    };
+
+    interface PolicyAuthorizationGroups {
+        logic: "AND" | "OR";
+        allowOperatorAsAuthorizer?: boolean;
+        groups: Array<AuthorizationGroup>;
+    }
+
+    export interface PolicyRule {
+        operator?: string;
+        operators?: {
+            wildcard?: "*";
+            users?: Array<string>;
+            usersGroups?: Array<string>;
+            services?: Array<string>;
+        };
+        transactionType?: PolicyTransactionType;
+        operatorServices?: Array<string>;
+        designatedSigner?: string;
+        designatedSigners?: {
+            users?: Array<string>;
+            usersGroups?: Array<string>;
+        };
+        type: PolicyType;
+        action: PolicyAction;
+        asset: string;
+        srcType?: PolicySrcOrDestType;
+        srcSubType?: PolicySrcOrDestSubType;
+        srcId?: PolicySrcOrDestId;
+        src?: {
+            ids?: Array<[PolicySrcOrDestId, PolicySrcOrDestType?, PolicySrcOrDestSubType?]>;
+        };
+        dstType?: PolicySrcOrDestType;
+        dstSubType?: PolicySrcOrDestSubType;
+        dstId?: PolicySrcOrDestId;
+        dst?: {
+            ids?: Array<[PolicySrcOrDestId, PolicySrcOrDestType?, PolicySrcOrDestSubType?]>;
+        };
+        dstAddressType?: PolicyDestAddressType;
+        amountCurrency: string;
+        amountScope: PolicyAmountScope;
+        amount: number | string;
+        periodSec: number;
+        authorizers?: Array<string>;
+        authorizersCount?: number;
+        authorizationGroups?: PolicyAuthorizationGroups;
+        amountAggregation?: {
+            operators: string;
+            srcTransferPeers: string;
+            dstTransferPeers: string;
+        };
+        rawMessageSigning?: {
+            derivationPath: {
+                path: Array<number>;
+            };
+            algorithm: string;
+        };
+        applyForApprove?: boolean;
+        applyForTypedMessage?: boolean;
+        externalDescriptor?: string;
+    }
+
+    interface Metadata {
+        editedBy?: string;
+        editedAt?: number;
+        publishedBy?: string;
+        publishedAt?: number;
+    }
+
+    enum PolicyStatus {
+        SUCCESS = "SUCCESS",
+        UNVALIDATED = "UNVALIDATED",
+        INVALID_CONFIGURATION = "INVALID_CONFIGURATION",
+        PENDING = "PENDING",
+        PENDING_CONSOLE_APPROVAL = "PENDING_CONSOLE_APPROVAL",
+        AWAITING_QUORUM = "AWAITING_QUORUM",
+        UNHANDLED_ERROR = "UNHANDLED_ERROR",
+    }
+
+    type PolicyBaseErrorField =
+        | "operator"
+        | "operators"
+        | "authorizationGroups"
+        | "designatedSigner"
+        | "designatedSigners"
+        | "contractMethods"
+        | "amountAggregation"
+        | "src"
+        | "dst"
+        | "";
+
+    interface PolicyRuleError {
+        errorMessage: string;
+        errorCodeName: string;
+        errorField: PolicyBaseErrorField;
+        errorCode: number;
+    }
+
+    interface PolicyRuleCheckResult {
+        index: number;
+        status: "ok" | "failure";
+        errors: Array<PolicyRuleError>;
+    }
+
+    interface PolicyCheckResult {
+        errors: number;
+        results: PolicyRuleCheckResult[];
+    }
+
+    interface PolicyResponse {
+        status?: string;
+        rules?: Array<PolicyRule>;
+        metadata?: Metadata;
+    }
+
+    interface ValidationResponse {
+        status?: PolicyStatus;
+        checkResult?: PolicyCheckResult;
+    }
+
+    interface DraftResponse {
+        draftId: string;
+        status?: PolicyStatus;
+        rules?: Array<PolicyRule>;
+        metadata?: Metadata;
+    }
+
+    export interface DraftReviewAndValidationResponse {
+        draftResponse: DraftResponse;
+        validation: ValidationResponse;
+    }
+
+    export interface PolicyAndValidationResponse {
+        policy: PolicyResponse;
+        validation: ValidationResponse;
+    }
+
+    export interface PublishResult {
+        status?: PolicyStatus;
+        rules?: Array<PolicyRule>;
+        checkResult?: PolicyCheckResult;
+        metadata?: Metadata;
     }
 }